src/streamlit_app/requirements.txt
altair==5.5.0
annotated-types==0.7.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
jellyfish==1.1.3
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.1
MarkupSafe==3.0.2
narwhals==1.31.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
proto-plus==1.26.1
protobuf==5.29.4
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pymongo==4.11.3
pyparsing==3.2.1
pypdf==5.4.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.23.1
rsa==4.9
segtok==1.5.11
six==1.17.0
smmap==5.0.2
streamlit==1.43.2
tabulate==0.9.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
watchdog==6.0.0
yake==0.4.8
